+++ libraries/OpenCL/ScalaCL2/src/main/scala/scalacl/ScalaCLUtils.scala
@@ -11,20 +11,20 @@ import org.bridj.SizeT
 object ScalaCLUtils {
 
   lazy val prefixSumCode = new CLSimpleCode("""
-    __kernel void prefSum(size_t size, __global const char* in, __global long* out) {
+    __kernel void prefSum(size_t size, __global const char* in, __global int* out) {
       size_t i = get_global_id(0);
       size_t j;
       if (i >= size)
         return;
 
-      long total = 0;
+      int total = 0;
       for (j = 0; j <= i; j++)
         total += in[j];
 
       out[i] = total;
     }
   """)
-  def prefixSum(bitmap: CLGuardedBuffer[Boolean], output: CLGuardedBuffer[Long])(implicit context: ScalaCLContext) = {
+  def prefixSum(bitmap: CLGuardedBuffer[Boolean], output: CLGuardedBuffer[Int])(implicit context: ScalaCLContext) = {
     val kernel = prefixSumCode.getKernel(context, bitmap, output)
     val globalSizes = Array(bitmap.size.asInstanceOf[Int])
     kernel.synchronized {
@@ -39,7 +39,7 @@ object ScalaCLUtils {
   lazy val copyPrefixedCode = new CLSimpleCode("""
     __kernel void copyPrefixed(
         size_t size,
-        __global const long* presencePrefix,
+        __global const int* presencePrefix,
         __global const char* in,
         size_t elementSize,
         __global char* out
@@ -48,7 +48,7 @@ object ScalaCLUtils {
       if (i >= size)
         return;
 
-      long prefix = presencePrefix[i];
+      int prefix = presencePrefix[i];
       if (!i && prefix > 0 || i && prefix > presencePrefix[i - 1]) {
         size_t j, inOffset = i * elementSize, outOffset = (prefix - 1) * elementSize;
         for (j = 0; j < elementSize; j++) {
@@ -57,7 +57,7 @@ object ScalaCLUtils {
       }
     }
   """)
-  def copyPrefixed[T](size: Long, presencePrefix: CLGuardedBuffer[Long], in: CLGuardedBuffer[T], out: CLGuardedBuffer[T])(implicit t: ClassManifest[T], context: ScalaCLContext) = {
+  def copyPrefixed[T](size: Int, presencePrefix: CLGuardedBuffer[Int], in: CLGuardedBuffer[T], out: CLGuardedBuffer[T])(implicit t: ClassManifest[T], context: ScalaCLContext) = {
     val kernel = copyPrefixedCode.getKernel(context, in, out)
     val globalSizes = Array(in.size.asInstanceOf[Int])
     val pio = PointerIO.getInstance(t.erasure)
